percentage now goes from 0.0 to 100.0.
w [weight] is global for updateWeights and randomInitialWeights.

k is global for deltaLearningRate.

calculatedOutputs is global for backpropagation.

crutial changes in derivatives. added modifiedTanh.

deltaLearningRate with a = learningRateIncrement and b = learningRateGeometricDecrement.

h and v global for computeOutputs. No activation applied for last layer.

deltaW global for computeAndAccumulateDeltas.

DONT try with batch -> backpropagationBatch not updated.

[!!!] -> The purpose of having variables global is to be able to change them from within other function (declaring them global inside them and clearing them afterwards outside).

Everything seems slow and very sensible to learningRate.